Ho fatto delle modifiche al tuo codice. Ho rimosso il driver e le proprieta per collegarsi a mysql e ho aggiunto H2 DB per fare dei test.

+++ filmlist/src/main/java/com/example/filmlist/repository/FilmRepository.java
@@ -3,9 +3,9 @@ package com.example.filmlist.repository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import com.exemple.filmlist.models.Film;
+import com.example.filmlist.entity.Film;
 
 @Repository
-public interface FilmRepository extends CrudRepository<Film, Integer> {
+public interface FilmRepository extends CrudRepository<Film, Long> {
 
 }
